+++ lib/src/dart/stream.dart
@@ -1,11 +1,38 @@
 import 'dart:async';
 
+import 'package:meta/meta.dart';
 import 'package:rxdart/rxdart.dart';
 // ignore: implementation_imports
 import 'package:rxdart/src/transformers/backpressure/backpressure.dart';
 
-extension StreamDartExt<T> on Stream<T> {
-  Stream<T> distinctType() => distinct((bef, aft) => bef.runtimeType == aft.runtimeType);
+extension StreamExtDart<T> on Stream<T> {
+  Stream<T> distinctRuntimeType() => distinct((bef, aft) => bef.runtimeType == aft.runtimeType);
+
+  Future<R> firstType<R>({R orElse()}) {
+    final completer = Completer<R>();
+    firstWhere((v) => v is T).then((v) => completer.complete(v as R), onError: (exception, stackTrace) {
+      if (exception is StateError && orElse != null) {
+        completer.complete(orElse());
+      } else {
+        completer.completeError(exception, stackTrace);
+      }
+    });
+    return completer.future;
+  }
+
+  Future<R> firstRuntimeType<R>({R orElse()}) {
+    final completer = Completer<R>();
+    firstWhere((v) => v.runtimeType == T).then((v) => completer.complete(v as R),
+        onError: (exception, stackTrace) {
+      if (exception is StateError && orElse != null) {
+        completer.complete(orElse());
+      } else {
+        completer.completeError(exception, stackTrace);
+      }
+    });
+    return completer.future;
+  }
+
   Stream<T> dumpErrorToConsoleDart() => doOnError((error, stackTrace) {
         print(error);
         print(stackTrace);
@@ -37,7 +64,79 @@ extension StreamDartExt<T> on Stream<T> {
         onWindowEnd: (queue) => queue.last,
         ignoreEmptyWindows: ignoreEmpty,
       ));
+
+  Stream<R> onChanges<R>({
+    bool Function(T, T) equals,
+    Duration debounceTime = const Duration(),
+    void Function(T data) onStart,
+    @required Stream<R> Function(T data) onData,
+  }) {
+    assert(debounceTime != null, 'debounceTime can\'t be null');
+    assert(onData != null, 'onData function is required');
+
+    final _onStart = onStart ?? (T d) {};
+
+    return distinct(equals).skip(1).doOnData(_onStart).debounceTime(debounceTime).switchMap<R>(onData);
+  }
 }
+
+extension StreamIterableExtDart<T> on Stream<Iterable<T>> {
+  Stream<R> onIterableChanges<R>({
+    bool Function(Iterable<T> previous, Iterable<T> current) equals,
+    Duration debounceTime = const Duration(),
+    void Function(Iterable<T> data) onStart,
+    @required Stream<R> Function(Iterable<T> previous, Iterable<T> current) onData,
+  }) {
+    assert(debounceTime != null, 'debounceTime can\'t be null');
+    assert(onData != null, 'onData function is required');
+
+    final _onStart = onStart ?? (T d) {};
+
+    return distinct(equals)
+        .pairwise()
+        .doOnData(_onStart)
+        .debounceTime(debounceTime)
+        .switchMap<R>((data) => onData(data.first, data.last));
+  }
+}
+//
+//@visibleForTesting
+//class SeparateStreamTransformer<T> extends StreamTransformerBase<Iterable<T>, SeparatedValues<T>> {
+//  SeparateStreamTransformer();
+//
+//  @override
+//  Stream<SeparatedValues<T>> bind(Stream<Iterable<T>> stream) async* {
+//    Iterable<T> oldValues = <T>[];
+//    await for (var currentValues in stream) {
+//      yield SeparatedValues<T>._(
+//        currentValues.where((value) => !oldValues.contains(value)).toList(),
+//        oldValues.where(currentValues.contains).toList(),
+//        oldValues.where((value) => !currentValues.contains(value)).toList(),
+//      );
+//      oldValues = currentValues;
+//    }
+//  }
+//}
+//
+//class SeparatedValues<T> {
+//  final List<T> newValues;
+//  final List<T> sameValues;
+//  final List<T> oldValues;
+//
+//  SeparatedValues._(this.newValues, this.sameValues, this.oldValues);
+//
+//  @override
+//  bool operator ==(Object other) =>
+//      identical(this, other) ||
+//      other is SeparatedValues &&
+//          runtimeType == other.runtimeType &&
+//          newValues == other.newValues &&
+//          sameValues == other.sameValues &&
+//          oldValues == other.oldValues;
+//
+//  @override
+//  int get hashCode => newValues.hashCode ^ sameValues.hashCode ^ oldValues.hashCode;
+//}
 
 extension FutureDartExt<T> on Future<T> {
   Future<T> dumpErrorToConsoleDart() => catchError((error, stackTrace) {
@@ -71,8 +170,7 @@ class CompositeMapSubscription {
 
   bool containsKey(Object key) => _subscriptions.containsKey(key);
 
-  bool containsSubscription(StreamSubscription subscription) =>
-      _subscriptions.containsValue(subscription);
+  bool containsSubscription(StreamSubscription subscription) => _subscriptions.containsValue(subscription);
 
   Future<dynamic> remove(Object key) => _subscriptions.remove(key)?.cancel();